+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.10] - DEV
+### Added
+- Add an option to toggle the dark theme.
+
 ### Changed
 - Improve look & feel of the menu.
 - Improve image filters of opening dialog.